test/installer: new layout, split unit/integration
Restructure the installer tests to follow the recent source restructure
