+++ lib/spack/spack/test/installer/__init__.py
@@ -1,0 +1,3 @@
+# Copyright Spack Project Developers. See COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: (Apache-2.0 OR MIT)
