+++ lib/spack/spack/test/installer/posix.py
