+++ lib/spack/spack/test/old_installer.py
@@ -18,7 +18,6 @@ import spack.database
 import spack.deptypes as dt
 import spack.error
 import spack.hooks
-import spack.installer_dispatch
 import spack.old_installer as inst
 import spack.package_base
 import spack.package_prefs as prefs
@@ -28,7 +27,6 @@ import spack.spec
 import spack.store
 import spack.util.filesystem as fs
 import spack.util.lock as lk
-from spack.main import SpackCommand
 from spack.test.conftest import RepoBuilder
 from spack.util import tty
 
@@ -660,88 +658,6 @@ def test_rewire_task_no_tarball(default_mock_concretization, monkeypatch):
     monkeypatch.setattr(spack.report.InstallRecord, "succeed", lambda x: None)
 
     assert rewire_task.complete() == inst.ExecuteResult.MISSING_BUILD_SPEC
-
-
-@pytest.mark.parametrize("transitive", [True, False])
-def test_install_spliced(install_mockery, mock_fetch, monkeypatch, transitive, installer_variant):
-    """Test installing a spliced spec"""
-    spec = spack.concretize.concretize_one("splice-t")
-    dep = spack.concretize.concretize_one("splice-h+foo")
-
-    # Do the splice.
-    out = spec.splice(dep, transitive)
-    installer = spack.installer_dispatch.create_installer(
-        [out.package], verbose=True, fail_fast=True
-    )
-    installer.install()
-    for node in out.traverse():
-        assert spack.store.STORE.db.installed(node)
-        assert spack.store.STORE.db.installed(node.build_spec)
-
-
-@pytest.mark.parametrize("transitive", [True, False])
-def test_install_spliced_build_spec_installed(
-    install_mockery, mock_fetch, transitive, installer_variant
-):
-    """Test installing a spliced spec with the build spec already installed"""
-    spec = spack.concretize.concretize_one("splice-t")
-    dep = spack.concretize.concretize_one("splice-h+foo")
-
-    # Do the splice.
-    out = spec.splice(dep, transitive)
-    spack.installer_dispatch.create_installer([out.build_spec.package]).install()
-
-    installer = spack.installer_dispatch.create_installer(
-        [out.package], verbose=True, fail_fast=True
-    )
-    installer.install()
-    for node in out.traverse():
-        assert spack.store.STORE.db.installed(node)
-        assert spack.store.STORE.db.installed(node.build_spec)
-
-
-# Unit tests should not be affected by the user's managed environments
-@pytest.mark.not_on_windows("lacking windows support for binary installs")
-@pytest.mark.parametrize("transitive", [True, False])
-@pytest.mark.parametrize(
-    "root_str", ["splice-t^splice-h~foo", "splice-h~foo", "splice-vt^splice-a"]
-)
-def test_install_splice_root_from_binary(
-    mutable_mock_env_path,
-    install_mockery,
-    mock_fetch,
-    temporary_mirror,
-    transitive,
-    root_str,
-    installer_variant,
-):
-    """Test installing a spliced spec with the root available in binary cache"""
-    # Test splicing and rewiring a spec with the same name, different hash.
-    original_spec = spack.concretize.concretize_one(root_str)
-    spec_to_splice = spack.concretize.concretize_one("splice-h+foo")
-
-    spack.installer_dispatch.create_installer(
-        [original_spec.package, spec_to_splice.package]
-    ).install()
-
-    out = original_spec.splice(spec_to_splice, transitive)
-
-    buildcache = SpackCommand("buildcache")
-    buildcache(
-        "push",
-        "--unsigned",
-        "--update-index",
-        temporary_mirror,
-        str(original_spec),
-        str(spec_to_splice),
-    )
-
-    uninstall = SpackCommand("uninstall")
-    uninstall("-ay")
-
-    spack.installer_dispatch.create_installer([out.package], unsigned=True).install()
-
-    assert len(spack.store.STORE.db.query()) == len(list(out.traverse()))
 
 
 class MockInstallStatus(inst.InstallStatus):
@@ -1402,11 +1318,148 @@ def test_print_install_test_log_failures(
     assert "See test results at" in out
 
 
-@pytest.mark.disable_clean_stage_check
-def test_log_files_preserved_on_error(install_mockery, mock_fetch, installer_variant):
-    """Test that the log file is preserved when an install error occurs."""
-    pkg = spack.concretize.concretize_one("build-error").package
-    installer = spack.installer_dispatch.create_installer([pkg])
-    with pytest.raises(spack.error.InstallError):
-        installer.install()
-    assert os.path.exists(pkg.log_path)
+def test_build_request_errors(install_mockery):
+    with pytest.raises(ValueError, match="must be a package"):
+        inst.BuildRequest("abc", {})
+
+    spec = spack.spec.Spec("trivial-install-test-package")
+    pkg_cls = spack.repo.PATH.get_pkg_class(spec.name)
+    with pytest.raises(ValueError, match="must have a concrete spec"):
+        inst.BuildRequest(pkg_cls(spec), {})
+
+
+def test_build_request_basics(install_mockery):
+    spec = spack.concretize.concretize_one("dependent-install")
+    assert spec.concrete
+
+    # Ensure key properties match expectations
+    request = inst.BuildRequest(spec.package, {})
+    assert not request.pkg.stop_before_phase
+    assert not request.pkg.last_phase
+    assert request.spec == spec.package.spec
+
+    # Ensure key default install arguments are set
+    assert "install_package" in request.install_args
+    assert "install_deps" in request.install_args
+
+
+def test_build_request_strings(install_mockery):
+    """Tests of BuildRequest repr and str for coverage purposes."""
+    # Using a package with one dependency
+    spec = spack.concretize.concretize_one("dependent-install")
+    assert spec.concrete
+
+    # Ensure key properties match expectations
+    request = inst.BuildRequest(spec.package, {})
+
+    # Cover __repr__
+    irep = request.__repr__()
+    assert irep.startswith(request.__class__.__name__)
+
+    # Cover __str__
+    istr = str(request)
+    assert "package=dependent-install" in istr
+    assert "install_args=" in istr
+
+
+@pytest.mark.parametrize(
+    "root_policy,dependencies_policy,package_deptypes,dependencies_deptypes",
+    [
+        ("auto", "auto", dt.BUILD | dt.LINK | dt.RUN, dt.BUILD | dt.LINK | dt.RUN),
+        ("cache_only", "auto", dt.LINK | dt.RUN, dt.BUILD | dt.LINK | dt.RUN),
+        ("auto", "cache_only", dt.BUILD | dt.LINK | dt.RUN, dt.LINK | dt.RUN),
+        ("cache_only", "cache_only", dt.LINK | dt.RUN, dt.LINK | dt.RUN),
+    ],
+)
+def test_build_request_deptypes(
+    install_mockery, root_policy, dependencies_policy, package_deptypes, dependencies_deptypes
+):
+    s = spack.concretize.concretize_one("dependent-install")
+
+    build_request = inst.BuildRequest(
+        s.package, {"root_policy": root_policy, "dependencies_policy": dependencies_policy}
+    )
+
+    actual_package_deptypes = build_request.get_depflags(s.package)
+    actual_dependency_deptypes = build_request.get_depflags(s["dependency-install"].package)
+
+    assert actual_package_deptypes == package_deptypes
+    assert actual_dependency_deptypes == dependencies_deptypes
+
+
+def test_build_task_errors(install_mockery):
+    """Check expected errors when instantiating a BuildTask."""
+    spec = spack.spec.Spec("trivial-install-test-package")
+    pkg_cls = spack.repo.PATH.get_pkg_class(spec.name)
+
+    # The value of the request argument is expected to not be checked.
+    for pkg in [None, "abc"]:
+        with pytest.raises(TypeError, match="must be a package"):
+            inst.BuildTask(pkg, None)
+
+    with pytest.raises(ValueError, match="must have a concrete spec"):
+        inst.BuildTask(pkg_cls(spec), None)
+
+    # Using a concretized package now means the request argument is checked.
+    spec = spack.concretize.concretize_one(spec)
+    assert spec.concrete
+
+    with pytest.raises(TypeError, match="is not a valid build request"):
+        inst.BuildTask(spec.package, None)
+
+    # Using a valid package and spec, the next check is the status argument.
+    request = inst.BuildRequest(spec.package, {})
+
+    with pytest.raises(TypeError, match="is not a valid build status"):
+        inst.BuildTask(spec.package, request, status="queued")
+
+    # Now we can check that build tasks cannot be create when the status
+    # indicates the task is/should've been removed.
+    with pytest.raises(spack.error.InstallError, match="Cannot create a task"):
+        inst.BuildTask(spec.package, request, status=inst.BuildStatus.REMOVED)
+
+    # Also make sure to not accept an incompatible installed argument value.
+    with pytest.raises(TypeError, match="'installed' be a 'set', not 'str'"):
+        inst.BuildTask(spec.package, request, installed="mpileaks")
+
+
+def test_build_task_basics(install_mockery):
+    spec = spack.concretize.concretize_one("dependent-install")
+    assert spec.concrete
+
+    # Ensure key properties match expectations
+    request = inst.BuildRequest(spec.package, {})
+    task = inst.BuildTask(spec.package, request=request, status=inst.BuildStatus.QUEUED)
+    assert not task.explicit
+    assert task.priority == len(task.uninstalled_deps)
+    assert task.key == (task.priority, task.sequence)
+
+    # Ensure flagging installed works as expected
+    assert len(task.uninstalled_deps) > 0
+    assert task.dependencies == task.uninstalled_deps
+    task.flag_installed(task.dependencies)
+    assert len(task.uninstalled_deps) == 0
+    assert task.priority == 0
+
+
+def test_build_task_strings(install_mockery):
+    """Tests of build_task repr and str for coverage purposes."""
+    # Using a package with one dependency
+    spec = spack.concretize.concretize_one("dependent-install")
+    assert spec.concrete
+
+    # Ensure key properties match expectations
+    request = inst.BuildRequest(spec.package, {})
+    task = inst.BuildTask(spec.package, request=request, status=inst.BuildStatus.QUEUED)
+
+    # Cover __repr__
+    irep = task.__repr__()
+    assert irep.startswith(task.__class__.__name__)
+    assert "BuildStatus.QUEUED" in irep
+    assert "sequence=" in irep
+
+    # Cover __str__
+    istr = str(task)
+    assert "status=queued" in istr  # == BuildStatus.QUEUED
+    assert "#dependencies=1" in istr
+    assert "priority=" in istr
